+++ src/components/autho0/Auth0History.tsx
@@ -2,26 +2,19 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import { Auth0Provider } from "@auth0/auth0-react";
 
-const Auth0ProviderWithHistory = ({ children }) => {
+const Auth0ProviderWorker = ({ children }) => {
   const domain = process.env.REACT_APP_AUTH0_DOMAIN;
   const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID;
-
-  const history = useNavigate();
-
-  const onRedirectCallback = (appState) => {
-    history.push(appState?.returnTo || window.location.pathname);
-  };
 
   return (
     <Auth0Provider
       domain={domain}
       clientId={clientId}
       redirectUri={window.location.origin}
-      onRedirectCallback={onRedirectCallback}
     >
       {children}
     </Auth0Provider>
   );
 };
 
-export default Auth0ProviderWithHistory;
+export default Auth0ProviderWorker;
