+++ src/index.tsx
@@ -12,7 +12,7 @@ import {
   cacheExchange,
   fetchExchange
 } from 'urql';
-import Auth0ProviderWithHistory from "./auth/auth0-provider-with-history";
+import Auth0ProviderWorker from "./components/autho0/Auth0History";
 
 const loadLocaleData = locale => {
   switch (locale) {
@@ -36,30 +36,13 @@ if (!window.__skip_render) {
 
   ReactDOM.render(
     <BrowserRouter>
-    <Auth0ProviderWithHistory>
-      <Provider value={client}>
-        <App locale={locale} messages={messages} />
-      </Provider>     
-    </Auth0ProviderWithHistory>
+      <Auth0ProviderWorker>
+        <Provider value={client}>
+          <App locale={locale} messages={messages} />
+        </Provider>     
+      </Auth0ProviderWorker>
     </BrowserRouter>,
     document.getElementById('root')
   );
 }
 serviceWorker.unregister();
-
-// import React from "react";
-// import ReactDOM from "react-dom";
-// import App from "./app";
-// import { BrowserRouter as Router } from "react-router-dom";
-// import Auth0ProviderWithHistory from "./auth/auth0-provider-with-history";
-
-// import "./index.css";
-
-// ReactDOM.render(
-//   <Router>
-//     <Auth0ProviderWithHistory>
-//       <App />
-//     </Auth0ProviderWithHistory>
-//   </Router>,
-//   document.getElementById("root")
-// );
